fix: glTF animation parse bug (#1845)
