+++ packages/loader/src/gltf/parser/GLTFAnimationParser.ts
@@ -40,7 +40,7 @@ export class GLTFAnimationParser extends GLTFParser {
     const entities = context.get<Entity>(GLTFParserType.Entity);
 
     let duration = -1;
-    let promises = new Array<Promise<void>>();
+    let promises = new Array<Promise<void | Entity[]>>();
 
     // parse samplers
     for (let j = 0, m = samplers.length; j < m; j++) {
@@ -95,6 +95,8 @@ export class GLTFAnimationParser extends GLTFParser {
       promises.push(promise);
     }
 
+    promises.push(context.get<Entity>(GLTFParserType.Scene));
+
     return Promise.all(promises).then(() => {
       for (let j = 0, m = channels.length; j < m; j++) {
         const gltfChannel = channels[j];
@@ -109,7 +111,7 @@ export class GLTFAnimationParser extends GLTFParser {
         }
 
         // If the target node is in the default scene, relativePath will be empty
-        if (entity !== context.glTFResource.defaultSceneRoot) {
+        if (context.glTFResource.sceneRoots.indexOf(entity) === -1) {
           continue;
         }
